pyproject.toml
[project]
name = "fastapi-auth-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "brevo-python>=1.2.0",
    "cloudinary>=1.44.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
